split date module
